+++ ayp2_plantilla_tp_grupal_2c2023-main/src/Inventario_de_Armas.cpp
@@ -122,7 +122,7 @@ void Inventario_de_Armas::desequipar_arma(){
 
 void Inventario_de_Armas::agregar_arma_aleatoria(){
     vector<string> vec_armas = {"Espada", "Arco", "Hacha", "Ballesta", "Daga", "Lanza", "Pistola", "Rifle", "Escopeta", "Subfusil", "Ametralladora"};
-    int indice_aleatorio = rand() % vec_armas.size();
+    int indice_aleatorio = rand() % (static_cast<int>(vec_armas.size()));
     string nombre_aleatorio = vec_armas[indice_aleatorio];
 
     size_t potencia_aleatoria = rand() % POTENCIA_MAX;
